+++ .ci/run_tests_github.sh
@@ -241,13 +241,14 @@ if [ "$TEST_SUITE" == "plugins-ee" ]; then
         # make test-degraphql || echo "* degraphql" >> .failed
         # make test-canary || echo "* canary" >> .failed
         # make test-opa || echo "* opa" >> .failed
-    
+
     elif [[ "$TEST_SPLIT" == third ]]; then
         make test-build-pongo-deps
         make test-mocking || echo "* mocking" >> .failed
         make test-tls-handshake-modifier || echo "* tls-handshake-modifier" >> .failed
         make test-upstream-timeout || echo "* upstream-timeout" >> .failed
         make test-key-auth-enc || echo "* key-auth-enc" >> .failed
+        make test-websocket-validator || echo "* websocket-validator" >> .failed
         # make test-rate-limiting-advanced || echo "* rate-limiting-advanced" >> .failed
         # make test-openid-connect || echo "* openid-connect" >> .failed
         # make test-route-transformer-advanced || echo "* route-transformer-advanced" >> .failed